Separate sample and texture lookup stores to avoid conflicts
As it turns out, current `master` can fail to look up a texture from
a skin, if the skin contains a texture and an audio file with the same
name, but different extensions.

The cause of this lay in how the resource stores in `Skin` were laid
out, and how the lookup process works. Two facts are relevant here:

- When looking up a file by name, a resource store will first try to
  treat the provided name as a full filename. If that fails and the
  store is associated with a particular set of extensions
  (via `AddExtension()`), then it will try appending those extensions,
  but only *after* the original lookup using the raw name fails.

- Resource stores nest, and perform lookups recursively. In the case
  of the texture store in `Skin`, the nesting was thus:

    MaxDimensionLimitedTextureLoaderStore
    TextureLoaderStore (created by host)
    Storage passed to `Skin` via ctor (or `RealmBackedResourceStore`, if none provided)

  However, in the constructor of `Skin`, the sample store is set up
  before the texture store. The setup of the sample store would end up
  calling `.AddExtension()` with audio extensions on the innermost
  storage.

Note that the fact that resource stores try the full filename first,
and only then try appending their extensions, means that if several
stores in a nested lookup have preferred extensions, then the priority
of these preferences is *inverted* (i.e. the innermost store wins).

Thus, because the innermost storage in this case was the one that ended
up having audio extensions tacked onto it, this meant that if both
an audio file and a texture with the same filename end up in a skin,
then the innermost storage would end up reading the audio file, but then
pass that to the texture loader which would try to interpret it as
a texture, and invariably fail.

In order to fix this, prevent the setup process of the sample store
from introducing extension bias towards audio files, by nesting the
storage passed to `Skin` one more level, and applying the extension
preference to that new nest level.