Bend beatmap encoder such that it treats numerical suffixes on non-legacy `HitSampleInfo`s like custom sample banks
This is dodgy as all hell, and at this point you might as well commit to
making `Suffix` a number. But I dunno. This is a difficult spot to crawl
out of.

If you profess the school of thought that "the stable way of samples is
completely terrible and should disappear" then `HitSampleInfo.Suffix`
shouldn't exist at all. But doing that requires a full skinning refactor
because the custom sample bank badness is smeared over beatmap skins and
you'd need *some* alternative facility to achieve the same results.

Maybe this is a half-measure. But for now this is as far as I'm willing
to go.