Fix custom sample bank 1 specs being discarded
The concept that this is implementing is quite bonkers to explain to the
user, which is why I'm not adding this to the UI in here because I have
to think about how this is to be presented even. But at least it fixes
the sample lookup test failures.

Lifting up the `UseBeatmapSamples` thing is unfortunate but I think
unavoidable given the extent to which people have come to expect these
behaviours.