Pass realm to beatmap working cache
This is a scary change but also one that makes sample reloading in
editor half-work.

The reason why it makes it half-work is that the beatmap skin goes
through `RealmBackedResourceStore` to perform lookups. If that isn't
given a realm, it will never invalidate the file cache that it
constructs, meaning it will be forever oblivious to file adds and
deletes. Passing realm in will make it not so.

Mind you, replacing sounds still doesn't work. But one step at a time...